+++ modules/markup/markdown/transform_heading.go
@@ -13,14 +13,14 @@ import (
 	"github.com/yuin/goldmark/text"
 )
 
-func (g *ASTTransformer) transformHeading(_ *markup.RenderContext, v *ast.Heading, reader text.Reader, tocList *[]markup.Header) {
+func (g *ASTTransformer) transformHeading(_ *markup.RenderContext, v *ast.Heading, reader text.Reader, tocList *[]Header) {
 	for _, attr := range v.Attributes() {
 		if _, ok := attr.Value.([]byte); !ok {
 			v.SetAttribute(attr.Name, []byte(fmt.Sprintf("%v", attr.Value)))
 		}
 	}
 	txt := v.Text(reader.Source()) //nolint:staticcheck
-	header := markup.Header{
+	header := Header{
 		Text:  util.UnsafeBytesToString(txt),
 		Level: v.Level,
 	}
