+++ modules/packages/debian/metadata_test.go
@@ -10,6 +10,7 @@ import (
 	"io"
 	"testing"
 
+	"code.gitea.io/gitea/modules/util"
 	"code.gitea.io/gitea/modules/zstd"
 
 	"github.com/blakesmith/ar"
@@ -77,7 +78,7 @@ func TestParsePackage(t *testing.T) {
 			{
 				Extension: "",
 				WriterFactory: func(w io.Writer) io.WriteCloser {
-					return nopCloser{w}
+					return util.NopCloser{Writer: w}
 				},
 			},
 			{
@@ -127,14 +128,6 @@ func TestParsePackage(t *testing.T) {
 			})
 		}
 	})
-}
-
-type nopCloser struct {
-	io.Writer
-}
-
-func (nopCloser) Close() error {
-	return nil
 }
 
 func TestParseControlFile(t *testing.T) {
