+++ modules/util/io.go
@@ -9,6 +9,12 @@ import (
 	"io"
 )
 
+type NopCloser struct {
+	io.Writer
+}
+
+func (NopCloser) Close() error { return nil }
+
 // ReadAtMost reads at most len(buf) bytes from r into buf.
 // It returns the number of bytes copied. n is only less than len(buf) if r provides fewer bytes.
 // If EOF or ErrUnexpectedEOF occurs while reading, err will be nil.
